+++ R/batch_writer.R
@@ -1,18 +1,19 @@
-#' Write an sbatch job file
+#' Write a slurm batch script file
 #'
-#' A wrapper around writing the script given the desired command, a short name for the job, a memory limit, and many optional parameters.
-#' Defaults are for my current setup (my email, latest R version on cluster).
+#' A wrapper around writing the script given the desired command, a short name for the job, and many optional parameters.
+#' Defaults are for my current setup (latest R version on cluster).
 #'
 #' @param commands A character vector of commands to run, one per line.
-#' @param name Short name for job, for queue file, and also for stdout/stderr output paths
+#' @param name Short name for for batch script file, job, and stdout/stderr output paths
+#' @param R Specify R version, to load module (`NA` prevents loading R).
 #' @param mem Memory limit.
+#' `NA` (default) is no memory limit.
 #' May be a string like `1G`, whatever sbatch accepts.
 #' @param time Time limit.
 #' `NA` (default) is no time limit.
 #' May be a string like `1:00:00`, whatever sbatch accepts.
 #' @param nodes Number of nodes to use (default 1).
 #' @param threads Number of tasks per node (default 1).
-#' @param R Specify R version, to load module (`NA` prevents loading R).
 #' @param email Address to receive emails about job completion.
 #'
 #' @examples
@@ -20,39 +21,36 @@
 #' # set desired parameters
 #' commands <- 'time echo "Hello world!"'
 #' name <- 'hola'
-#' mem <- '1G'
 #' 
 #' # writes file `hola.q`
 #' batch_writer(
 #'     commands,
-#'     name,
-#'     mem
+#'     name
 #' )
 #'
 #' # cleanup
-#' file.remove( paste0( name, '.q' ) )
+#' batch_cleanup( name )
 #' 
 #' @seealso
-#' `\link{batch_submit}` to submit these batch files
+#' `\link{batch_submit}` to submit slurm batch scripts.
+#' `\link{batch_cleanup}` to remove slurm batch scripts.
 #'
 #' @export
 batch_writer <- function(
                          commands,
                          name,
-                         mem,
+                         R = '4.0.0',
+                         mem = NA,
                          time = NA,
                          nodes = 1,
                          threads = 1,
-                         R = '4.0.0',
-                         email = 'alejandro.ochoa@duke.edu'
+                         email = NA
                          ) {
     # make sure nothing important is missing
     if ( missing( commands ) )
         stop('`commands` is required!')
     if ( missing( name ) )
         stop('`name` is required!')
-    if ( missing( mem ) )
-        stop('`mem` is required!')
 
     # wrap commands with load/unload R (common option!)
     if ( !is.na( R ) ) {
@@ -64,22 +62,60 @@ batch_writer <- function(
     }
 
     # add repetitive/structured sbatch header lines
-    commands <- c(
+    # these are always present
+    header <- c(
         '#!/bin/bash',
         paste0( '#SBATCH --job-name=', name ),
-        paste0( '#SBATCH --mail-user=', email ),
-        '#SBATCH --mail-type=END,FAIL',
-        paste0( '#SBATCH --mem=', mem ),
-        if ( !is.na(time) ) paste0( '#SBATCH --time=', time ) else '',
-        paste0( '#SBATCH --nodes=', nodes ),
-        paste0( '#SBATCH --ntasks-per-node=', threads ),
-        paste0( '#SBATCH --output=', name, '.stdout' ),
-        paste0( '#SBATCH --error=', name, '.stderr' ),
+        paste0( '#SBATCH --output=', name, '.out' )
+        ## paste0( '#SBATCH --error=', name, '.stderr' )
+    )
+    
+    # add email info if provided
+    if ( !is.na( email ) )
+        header <- c(
+            header,
+            paste0( '#SBATCH --mail-user=', email ),
+            '#SBATCH --mail-type=END,FAIL'
+        )
+    
+    # add memory limit if provided
+    if ( !is.na( mem ) )
+        header <- c(
+            header,
+            paste0( '#SBATCH --mem=', mem )
+        )
+        
+    # add time limit if provided
+    if ( !is.na( time ) )
+        header <- c(
+            header,
+            paste0( '#SBATCH --time=', time )
+        )
+
+    # add nodes option if greater than 1
+    if ( nodes > 1 )
+        header <- c(
+            header,
+            paste0( '#SBATCH --nodes=', nodes )
+        )
+
+    # add "threads" if greater than 1
+    if ( threads > 1 )
+        header <- c(
+            header,
+            paste0( '#SBATCH --ntasks-per-node=', threads )
+        )
+    
+    # merge header and commands
+    commands <- c(
+        header,
         commands
     )
 
-    # write to a file
+    # batch file full path
     file <- paste0( name, '.q' )
+    
+    # write script to file
     readr::write_lines(
         x = commands,
         path = file